+++ client/package.json
@@ -3,9 +3,12 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@date-io/date-fns": "^2.11.0",
     "@material-ui/core": "^4.11.0",
     "@material-ui/icons": "^4.11.2",
-    "@material-ui/lab": "^4.0.0-alpha.58",
+    "@material-ui/lab": "5.0.0-alpha.24",
+    "@mui/icons-material": "^5.1.1",
+    "@mui/lab": "^5.0.0-alpha.55",
     "@testing-library/jest-dom": "^5.11.9",
     "@testing-library/react": "^11.2.5",
     "@testing-library/user-event": "^13.0.6",
@@ -15,13 +18,16 @@
     "@types/react-dom": "^17.0.2",
     "body-parser": "^1.19.0",
     "clsx": "^1.1.1",
+    "datetimepicker": "^0.1.39",
     "env-cmd": "^10.1.0",
     "eslint": "^7.26.0",
     "formik": "^2.2.6",
     "immer": "^9.0.6",
+    "ngx-mat-datefns-date-adapter": "^11.2.2",
     "react": "^17.0.1",
     "react-beautiful-dnd": "^13.1.0",
     "react-dom": "^17.0.1",
+    "react-localization": "^1.0.17",
     "react-router-dom": "^5.2.0",
     "react-scripts": "^4.0.0",
     "socket.io-client": "^4.0.1",
