+++ client/src/components/Kanban/Card/CardDetails/CardDetailsComponents/CardComment.tsx
@@ -1,43 +1,42 @@
 import { IconButton } from '@mui/material';
-import AccessTimeOutlinedIcon from '@mui/icons-material/AccessTimeOutlined';
+import ModeCommentOutlinedIcon from '@mui/icons-material/ModeCommentOutlined';
 import useStyles from '../../useStyles';
 import CloseIcon from '@mui/icons-material/Close';
-import { Box, CircularProgress, Typography } from '@material-ui/core';
+import { useEffect, useState } from 'react';
+import { Box, Typography, TextareaAutosize, CircularProgress } from '@material-ui/core';
 import Button from '@mui/material/Button';
-import TextField from '@mui/material/TextField';
-import * as React from 'react';
-import { IIds } from '../../../../interface/Boards';
 import DialogContentText from '@mui/material/DialogContentText';
-import { saveCardItem, deleteCardItem } from '../../../../helpers/APICalls/cardApiCalls';
-import { useSnackBar } from '../../../../context/useSnackbarContext';
-import { useState, useEffect } from 'react';
-import { useBoard } from '../../../../context/useBoardContext';
+import { saveCardItem, deleteCardItem } from '../../../../../helpers/APICalls/cardApiCalls';
+import { IIds } from '../../../../../interface/Board';
+import { useSnackBar } from '../../../../../context/useSnackbarContext';
+import { useBoard } from '../../../../../context/useBoardContext';
 
 interface Props {
-  deadline: string | undefined;
+  comment: string | undefined;
   disableSetting: boolean;
   ids: IIds | undefined;
 }
 
-export default function CardDeadline({ deadline, disableSetting, ids }: Props): JSX.Element {
+export default function CardComment({ comment, disableSetting, ids }: Props): JSX.Element {
   const classes = useStyles();
-  const [date, setDate] = useState(deadline);
+  const [newComment, setNewComment] = useState(comment);
   const { updateSnackBarMessage } = useSnackBar();
   const [isSubmitting, setIsSubmitting] = useState(false);
-  const { updateBoards, boards } = useBoard();
+  const { updateBoard } = useBoard();
 
   useEffect(() => {
-    setDate(deadline);
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [boards]);
+    if (comment) {
+      setNewComment(comment);
+    }
+  }, [comment]);
 
-  const handleSaveDeadline = async (): Promise<void> => {
+  const handleSaveComment = async (): Promise<void> => {
     setIsSubmitting(true);
-    saveCardItem('deadline', date, ids).then((data) => {
+    saveCardItem('comment', newComment, ids).then((data) => {
       if (data.error) {
         updateSnackBarMessage(data.error.message);
       } else if (data.success) {
-        updateBoards();
+        updateBoard();
       } else {
         // should not get here from backend but this catch is for an unknown issue
         console.error({ data });
@@ -47,13 +46,13 @@ export default function CardDeadline({ deadline, disableSetting, ids }: Props): 
     });
     setIsSubmitting(false);
   };
-  const handleDeleteDeadline = async (): Promise<void> => {
+  const handleDeleteComment = async (): Promise<void> => {
     setIsSubmitting(true);
-
-    deleteCardItem('deadline', date, ids).then((data) => {
+    deleteCardItem('comment', newComment, ids).then((data) => {
       if (data.error) {
         updateSnackBarMessage(data.error.message);
       } else if (data.success) {
+        updateBoard();
       } else {
         // should not get here from backend but this catch is for an unknown issue
         console.error({ data });
@@ -68,45 +67,36 @@ export default function CardDeadline({ deadline, disableSetting, ids }: Props): 
     <>
       <Box className={classes.dialogContentTextContainer}>
         <DialogContentText className={classes.dialogContentText}>
-          <AccessTimeOutlinedIcon className={classes.textSpacing} />
+          <ModeCommentOutlinedIcon className={classes.textSpacing} />
           <Typography className={classes.cardItemTitle} variant="h6">
-            {' '}
-            Deadline:
+            Add Comment:
           </Typography>
         </DialogContentText>
         <Box className={classes.textAreaContainer}>
-          <Box>
-            {' '}
-            <TextField
-              id="datetime-local"
-              type="datetime-local"
-              defaultValue={date}
-              onChange={(e) => setDate(e.target.value)}
-              sx={{ width: 250 }}
-              disabled={disableSetting}
-              InputLabelProps={{
-                shrink: true,
-              }}
-              variant="standard"
-            />
-          </Box>
+          <TextareaAutosize
+            value={newComment}
+            aria-label="minimum height"
+            placeholder="There is no description for this card"
+            className={classes.textArea}
+            disabled={disableSetting}
+            onChange={(e) => setNewComment(e.target.value)}
+          />
           <Box className={classes.saveButtonContainer}>
             {' '}
             <Button
-              variant="contained"
-              onClick={handleSaveDeadline}
+              onClick={handleSaveComment}
               disabled={disableSetting || isSubmitting}
               size="small"
+              variant="contained"
             >
               {isSubmitting ? <CircularProgress style={{ fontSize: 0, width: '20px', height: '20px' }} /> : 'Save'}
             </Button>{' '}
             <IconButton
-              aria-label="close"
-              color="primary"
               disabled={disableSetting}
-              style={{ fontSize: '20px' }}
-              onClick={handleDeleteDeadline}
+              color="primary"
+              aria-label="close"
               size="small"
+              onClick={handleDeleteComment}
             >
               <CloseIcon />
             </IconButton>
