+++ client/src/components/Kanban/Card/CardDetails/CardDetailsComponents/CardDescription.tsx
@@ -1,57 +1,42 @@
 import { IconButton } from '@mui/material';
-import ModeCommentOutlinedIcon from '@mui/icons-material/ModeCommentOutlined';
+import DescriptionOutlinedIcon from '@mui/icons-material/DescriptionOutlined';
 import useStyles from '../../useStyles';
 import CloseIcon from '@mui/icons-material/Close';
 import { useEffect, useState } from 'react';
-import { Box, Typography, TextareaAutosize, CircularProgress } from '@material-ui/core';
+import { Box, Typography, CircularProgress, TextareaAutosize } from '@material-ui/core';
 import Button from '@mui/material/Button';
 import DialogContentText from '@mui/material/DialogContentText';
-import { saveCardItem, deleteCardItem } from '../../../../helpers/APICalls/cardApiCalls';
-import { IIds } from '../../../../interface/Boards';
-import { useSnackBar } from '../../../../context/useSnackbarContext';
-import { useBoard } from '../../../../context/useBoardContext';
+import { saveCardItem, deleteCardItem } from '../../../../../helpers/APICalls/cardApiCalls';
+import { IIds } from '../../../../../interface/Board';
+import { useSnackBar } from '../../../../../context/useSnackbarContext';
+import { useBoard } from '../../../../../context/useBoardContext';
 
 interface Props {
-  comment: string | undefined;
+  description: string | undefined;
   disableSetting: boolean;
   ids: IIds | undefined;
 }
 
-export default function CardComment({ comment, disableSetting, ids }: Props): JSX.Element {
+export default function CardDescription({ description, disableSetting, ids }: Props): JSX.Element {
   const classes = useStyles();
-  const [newComment, setNewComment] = useState(comment);
+  const [newDescription, setNewDescription] = useState(description);
   const { updateSnackBarMessage } = useSnackBar();
   const [isSubmitting, setIsSubmitting] = useState(false);
-  const { updateBoards, boards } = useBoard();
+  const { updateBoard } = useBoard();
 
   useEffect(() => {
-    setNewComment(comment);
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [boards]);
+    if (description) {
+      setNewDescription(description);
+    }
+  }, [description]);
 
-  const handleSaveComment = async (): Promise<void> => {
-    setIsSubmitting(true);
-    saveCardItem('comment', newComment, ids).then((data) => {
+  const handleSaveDescription = async (): Promise<void> => {
+    saveCardItem('description', newDescription, ids).then((data) => {
+      setIsSubmitting(true);
       if (data.error) {
         updateSnackBarMessage(data.error.message);
       } else if (data.success) {
-        updateBoards();
-      } else {
-        // should not get here from backend but this catch is for an unknown issue
-        console.error({ data });
-
-        updateSnackBarMessage('An unexpected error occurred. Please try again');
-      }
-    });
-    setIsSubmitting(false);
-  };
-  const handleDeleteComment = async (): Promise<void> => {
-    setIsSubmitting(true);
-    deleteCardItem('comment', newComment, ids).then((data) => {
-      if (data.error) {
-        updateSnackBarMessage(data.error.message);
-      } else if (data.success) {
-        updateBoards();
+        updateBoard();
       } else {
         // should not get here from backend but this catch is for an unknown issue
         console.error({ data });
@@ -62,40 +47,60 @@ export default function CardComment({ comment, disableSetting, ids }: Props): JS
     setIsSubmitting(false);
   };
 
+  const handleDeleteDescription = async (): Promise<void> => {
+    deleteCardItem('description', newDescription, ids).then((data) => {
+      setIsSubmitting(true);
+      if (data.error) {
+        updateSnackBarMessage(data.error.message);
+      } else if (data.success) {
+        // updateCardContext(data.success);
+      } else {
+        // should not get here from backend but this catch is for an unknown issue
+        console.error({ data });
+
+        updateSnackBarMessage('An unexpected error occurred. Please try again');
+      }
+    });
+    setIsSubmitting(false);
+  };
   return (
     <>
       <Box className={classes.dialogContentTextContainer}>
         <DialogContentText className={classes.dialogContentText}>
-          <ModeCommentOutlinedIcon className={classes.textSpacing} />
+          <DescriptionOutlinedIcon className={classes.textSpacing} />
           <Typography className={classes.cardItemTitle} variant="h6">
-            Add Comment:
+            {' '}
+            Description:
           </Typography>
         </DialogContentText>
         <Box className={classes.textAreaContainer}>
           <TextareaAutosize
-            value={newComment}
+            value={newDescription}
             aria-label="minimum height"
-            placeholder="There is no description for this card"
+            onChange={(e) => setNewDescription(e.target.value)}
+            placeholder="This Card has no Description "
             className={classes.textArea}
+            style={{ minHeight: '5rem' }}
             disabled={disableSetting}
-            onChange={(e) => setNewComment(e.target.value)}
           />
+
           <Box className={classes.saveButtonContainer}>
             {' '}
             <Button
-              onClick={handleSaveComment}
+              onClick={handleSaveDescription}
               disabled={disableSetting || isSubmitting}
-              size="small"
               variant="contained"
+              size="small"
             >
               {isSubmitting ? <CircularProgress style={{ fontSize: 0, width: '20px', height: '20px' }} /> : 'Save'}
             </Button>{' '}
             <IconButton
               disabled={disableSetting}
-              color="primary"
               aria-label="close"
+              style={{ fontSize: '20px' }}
+              color="primary"
               size="small"
-              onClick={handleDeleteComment}
+              onClick={handleDeleteDescription}
             >
               <CloseIcon />
             </IconButton>
