+++ client/src/components/Kanban/Card/CardDetails/CardDetailsComponents/CardTitle.tsx
@@ -2,36 +2,36 @@ import useStyles from '../../useStyles';
 import { useEffect, useState } from 'react';
 import { Typography } from '@material-ui/core';
 import { TextField } from '@mui/material';
-import { saveCardItem } from '../../../../helpers/APICalls/cardApiCalls';
-import { IIds } from '../../../../interface/Boards';
-import { useSnackBar } from '../../../../context/useSnackbarContext';
-import { useBoard } from '../../../../context/useBoardContext';
+import { saveCardItem } from '../../../../../helpers/APICalls/cardApiCalls';
+import { IIds } from '../../../../../interface/Board';
+import { useSnackBar } from '../../../../../context/useSnackbarContext';
+import { useBoard } from '../../../../../context/useBoardContext';
 
 interface Props {
-  cardTitle: string | undefined;
+  name: string | undefined;
   disableSetting: boolean;
   ids: IIds | undefined;
 }
 
-export default function CardTitle({ cardTitle, disableSetting, ids }: Props): JSX.Element {
+export default function CardTitle({ name, disableSetting, ids }: Props): JSX.Element {
   const classes = useStyles();
   const { updateSnackBarMessage } = useSnackBar();
-  const [title, setTitle] = useState(cardTitle);
+  const [title, setTitle] = useState(name);
   const [enableTitleEdit, setEnableTitleEdit] = useState(false);
-  const { updateBoards, boards } = useBoard();
+  const { updateBoard } = useBoard();
 
   useEffect(() => {
-    setTitle(cardTitle);
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [boards]);
-
+    if (name) {
+      setTitle(name);
+    }
+  }, [name]);
   const handleSaveTitle = async (): Promise<void> => {
     setEnableTitleEdit(false);
     saveCardItem('title', title, ids).then((data) => {
       if (data.error) {
         updateSnackBarMessage(data.error.message);
       } else if (data.success) {
-        updateBoards();
+        updateBoard();
       } else {
         // should not get here from backend but this catch is for an unknown issue
         console.error({ data });
