+++ client/src/interface/Board.ts
@@ -1,5 +1,6 @@
 import { Column } from './Column';
 import { Card } from './Card';
+import { ICard } from './Card';
 
 export interface Board {
   id: string;
@@ -8,4 +9,20 @@ export interface Board {
   cards: Array<Card>;
   user: string;
   createdAt: string;
+}
+
+export interface IBoard {
+  id: string;
+  boardTitle: string;
+  columns: {
+    id: string;
+    columnTitle: string;
+    cards: ICard[];
+  }[];
+}
+
+export interface IIds {
+  cardId: string;
+  columnId: string;
+  boardId: string;
 }
