+++ client/src/interface/Card.ts
@@ -6,4 +6,24 @@ export interface Card {
   dueDate?: Date;
   description?: string;
   createdAt?: Date;
+  comment?: string;
+}
+
+export interface ICard {
+  id: string;
+  name?: string;
+  tag?: string;
+  description?: string;
+  comment?: string;
+  dueDate?: Date;
+  checklist?: {
+    item: string;
+    isChecked: boolean;
+    id: string;
+  }[];
+  attachment?: {
+    imageName: string;
+    imageSource: string;
+    id: string;
+  };
 }
